+++ helpers/webhooks/getGuildWebhooks.ts
@@ -15,7 +15,7 @@ import { Collection } from "../../util/collection.ts";
  *
  * @see {@link https://discord.com/developers/docs/resources/webhook#get-guild-webhooks}
  */
-export async function getWebhooks(bot: Bot, guildId: bigint): Promise<Collection<bigint, Webhook>> {
+export async function getGuildWebhooks(bot: Bot, guildId: bigint): Promise<Collection<bigint, Webhook>> {
   const results = await bot.rest.runMethod<DiscordWebhook[]>(
     bot.rest,
     "GET",
