+++ helpers/webhooks/mod.ts
@@ -7,7 +7,8 @@ export * from "./editWebhook.ts";
 export * from "./editWebhookMessage.ts";
 export * from "./editWebhookWithToken.ts";
 export * from "./executeWebhook.ts";
+export * from "./getChannelWebhooks.ts";
+export * from "./getGuildWebhooks.ts";
 export * from "./getWebhook.ts";
 export * from "./getWebhookMessage.ts";
-export * from "./getWebhooks.ts";
 export * from "./getWebhookWithToken.ts";
